Remove the intercept assumption in Gibbs samplers
The samplers currently assume that there is an intercept term in the
regressions, and this term is treated separately from the other
regression parameters (Makalic & Schmidt 2016). This is limiting for two
reasons:

1. We don't always want to run regressions with an intercept parameter;
at the very least we'd like to control its prior value.
2. This complicates the work for #16.

As a result in this commit we generalize the Gibbs samplers by removing
the separation between the intercept beta0 and the other regression parameters.