Build samplers using the model graph as an input
The current sampler constructors are not easily generalizable as they
take as inputs parameters that are specific to the model the posterior
distribution of which we want to sample from.

Therefore in this PR we standardize the sampler constructors which now
both take as input the model graph, the observed data and the number of
samples. We use simple pattern matching to check that the input model
matches the model that the sampler is designed to sample from.